requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
httpx==0.28.1
structlog==24.4.0
python-dotenv==1.0.1
pyyaml==6.0.2
faster-whisper==1.1.0
numpy==2.4.4
websockets==13.1
psutil==7.2.2
nvidia-ml-py==13.595.45
python-multipart==0.0.24
ai-edge-litert==2.1.4
pillow==12.2.0
scipy==1.15.3
pygltflib==1.16.3
